fix(protocol): break the runtime import cycle the capability barrels introduced
CI caught a ReferenceError from a temporal dead zone: collapsing to one barrel
per capability converted narrow cross-capability edges into wide ones, and the
composition root closed a loop.

    shared/agent/tool.factory.ts
      -> participant-agents/index.ts   (re-exports the chat personas)
      -> chat/*.persona.ts             (import createChatTools)
      -> shared/agent/tool.factory.ts

The old facade was narrow and pointed straight at agent.tools.js, so the loop
did not exist. The composition root is exempt from the barrel rule precisely
because it must reach every capability, so it now imports the leaf directly.

Two leaf contracts also moved out of the capabilities that happened to host
them, into the neutral layer both sides already depend on:

- opportunity/domain/opportunity.claim-safety.ts -> shared/utils/claim-safety.ts.
  Three pure text predicates with no imports of their own. negotiation and
  contacts needed them and were pulling the whole opportunity capability, which
  is what broke contact.inviter's partial mock of model.config.
- UnderspecificationTypeSchema -> shared/schemas/underspecification.schema.ts.
  The signals clarifier was importing the entire questions capability, LLM
  agents and tools included, to reach a three-value enum.

Both are shared contracts rather than capability internals, so filing them
neutrally removes the coupling instead of routing around it.

Also worth recording: the first measurement of this was wrong. The detector
counted `import type` edges, which TypeScript erases, and reported an 87-file
cycle. Counting only runtime edges, dev has 0 and this branch had 1 cycle of 6
files. The regression was real but small and local, not architectural.

Verification (bun scripts/test.ts is the isolated per-file runner CI uses; the
shared-process `bun test src` masked this entirely):
- bun scripts/test.ts: 2218 pass, 0 fail, 0 errors (dev: 2219 pass, 0 fail)
- runtime import cycles: 0, matching dev
- public surface: 443 exported symbols byte-identical to dev
- bun run architecture:check: exit 0
- bun run check:protocol-atlas: artifact current
- services/api tsc --noEmit: clean
