+++ packages/protocol/src/contacts/application/contact.inviter.ts
@@ -13,7 +13,7 @@ import { HumanMessage, SystemMessage } from "@langchain/core/messages";
 import { z } from "zod";
 
 import { createStructuredModel } from "../../shared/agent/model.config.js";
-import { stripUnsupportedOpportunityClaims } from "../../opportunity/index.js";
+import { stripUnsupportedOpportunityClaims } from "../../shared/utils/claim-safety.js";
 
 const InviteInputSchema = z.object({
   recipientName: z.string(),
