+++ packages/protocol/src/negotiation/domain/negotiation.question-safety.ts
@@ -1,4 +1,4 @@
-import { hasUnsupportedOpportunityClaim } from '../../opportunity/index.js';
+import { hasUnsupportedOpportunityClaim } from '../../shared/utils/claim-safety.js';
 
 /** Fixed prompt-safe labels; producers must not replace them with raw network/counterparty text. */
 export const NEGOTIATION_QUESTION_GENERIC_COUNTERPARTY = 'the other participant';
