+++ packages/protocol/src/shared/utils/claim-safety.ts
